techhire/requirements.txt
Django>=4.2,<7.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-filter>=23.0
